Add support for supplying initial state through constructor
Currently there is an issue where the 'on change' observer would be
called when setting the initial state of the history. This may not
necessarily be ideal. Right now the only way to avoid this is to
manually reassign 'history.onChange' after constructing the History
object. This feels quite hacky, and isn't supported by documentation.

By passing the initial state of the data into the constructor, a
documented way of handling this situation is provided. Critically, it
avoids the need to be fully aware of the internals of the History class.